+++ packages/metal-core/src/command/index.ts
@@ -1,5 +1,6 @@
 export * from "./types";
 export * from "./aliases";
 export * from "./parser";
+export * from "./line";
 export * from "./suggest";
 export * from "./format";